refactor: remove autosave functionality from `EnabledMods`
This feature is nice in theory but in practice can be inconsistent and can very easily silently fail, leading to potentially hard to find bugs.
It's much easier to tell what's going on with an explicit `save` call, so better to just force that than leave a magic footgun laying around.